requirements.txt
coverage==6.5.0
django-cors-headers==3.13.0
Django==3.2.14
djangorestframework==3.14.0
psycopg2-binary==2.9.5
drf-yasg==1.21.4
gunicorn==20.0.4
django-classy-settings==3.0.1
django-filter==22.1
django-oauth-toolkit==2.2.0  
